Fixes badly  imported structure
SP_DEVICE_INTERFACE_DATA structure is defind with a platform specific ULONG_PTR which in C# is IntPtr not int. This was causing a 1784 system error when calling SetupDiEnumDeviceInterfaces in 64bit builds.

(There maybe more inconsistencies in the imported structures)

+++ USB/Structures/SpDeviceInterfaceData.cs
@@ -3,12 +3,13 @@ using System.Runtime.InteropServices;
 
 namespace UsbHid.USB.Structures
 {
+    /// <see cref="https://msdn.microsoft.com/en-us/library/windows/hardware/ff552342(v=vs.85).aspx"/>
     [StructLayout(LayoutKind.Sequential)]
     public struct SpDeviceInterfaceData
     {
-        public int cbSize;
+        public Int32 cbSize;
         public Guid InterfaceClassGuid;
-        public int Flags;
-        public int Reserved;
+        public Int32 Flags;
+        public IntPtr Reserved;
     }
 }
